(Misc. Cleanup) Remove references to secdec
