+++ src/main/java/com/codedx/api/client/GetProjectsResponse.java
@@ -12,34 +12,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License
  */
-package com.secdec.codedx.api.client;
+package com.codedx.api.client;
+
+import java.util.List;
 
 /**
- * Represents the JSON data for a Job
+ * Used to wrap the list of projects returned by the getProjects call.
+ * 
+ * I guess eventually pagination-related properties would go here.
  * 
  * @author anthonyd
  *
  */
-public class Job {
+public class GetProjectsResponse {
 
-	private String jobId;
-	private String status;
-	
-	public static final String QUEUED = "queued";
-	public static final String RUNNING = "running";
-	public static final String COMPLETED = "completed";
-	public static final String FAILED = "failed";
-	
-	public String getJobId() {
-		return jobId;
+	private List<Project> projects;
+
+	public List<Project> getProjects() {
+		return projects;
 	}
-	public void setJobId(String jobId) {
-		this.jobId = jobId;
-	}
-	public String getStatus() {
-		return status;
-	}
-	public void setStatus(String status) {
-		this.status = status;
+
+	public void setProjects(List<Project> projects) {
+		this.projects = projects;
 	}
 }
