+++ src/main/java/com/codedx/security/CompositeX509TrustManager.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License
  */
-package com.secdec.codedx.security;
+package com.codedx.security;
 
 import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
