+++ src/main/java/com/codedx/security/X509HostnameVerifierWithExceptions.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License
  */
-package com.secdec.codedx.security;
+package com.codedx.security;
 
 import java.util.Set;
