+++ src/main/java/com/codedx/util/HashUtil.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License
  */
-package com.secdec.codedx.util;
+package com.codedx.util;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
